requirements.txt
anyio==4.2.0
APScheduler==3.10.4
certifi==2023.11.17
dnspython==2.4.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
log4mongo==1.7.0
paho-mqtt==1.6.1
pymongo==4.5.0
python-dateutil==2.8.2
python-telegram-bot==20.7
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
suntime==1.2.5
tzdata==2023.3
tzlocal==5.2
validators==0.22.0
